Refactor module structure: Separate Ruler and Workflow responsibilities
- Restructure modules with clear separation of responsibilities
- Move Ruler to focus on rule & agent management only
- Create new Workflow module for workflow management & execution
- Remove rule_engine directory and consolidate components
- Move session and hot_reload to workflow module
- Consolidate rule_file.rs + compiled_rule.rs into rule_types.rs + rule_loader.rs
- Update all import statements throughout codebase
- All tests pass and code quality checks succeed
