+++ src/ruler/mod.rs
@@ -1,27 +1,28 @@
-pub mod rule_engine;
-pub mod session;
+pub mod decision;
+pub mod rule_loader;
+pub mod rule_types;
 
 use crate::agent::Agent;
-use crate::ruler::rule_engine::{decide_action, ActionType, CmdKind, RuleEngine};
-use crate::ruler::session::SessionStore;
+use crate::ruler::decision::decide_action;
+use crate::ruler::rule_types::{ActionType, CompiledRule};
 use anyhow::Result;
 use std::collections::HashMap;
-use std::path::PathBuf;
 use std::sync::Arc;
-use tokio::sync::Mutex;
-use tokio::time::Duration;
+use tokio::sync::RwLock;
 
 pub struct Ruler {
-    rule_engine: Arc<RuleEngine>,
+    rules: Arc<RwLock<Vec<CompiledRule>>>,
     agents: HashMap<String, Agent>,
-    sessions: Arc<Mutex<SessionStore>>,
     test_mode: bool,
 }
 
 #[allow(dead_code)]
 impl Ruler {
     pub async fn new(rules_path: &str) -> Result<Self> {
-        let rule_engine = RuleEngine::new(rules_path).await?;
+        // Load initial rules
+        let initial_rules =
+            crate::ruler::rule_loader::load_rules(std::path::Path::new(rules_path))?;
+        let rules = Arc::new(RwLock::new(initial_rules));
 
         // In test environment, create a simple mock backend that always succeeds
         let is_test = std::env::var("CARGO_TEST").is_ok()
@@ -41,25 +42,9 @@ impl Ruler {
                 })
                 .unwrap_or(false);
 
-        // Initialize session store with default persistence file
-        let session_file = if is_test {
-            PathBuf::from("/tmp/rule-agents-test-sessions.json")
-        } else {
-            dirs::config_dir()
-                .unwrap_or_else(|| PathBuf::from("."))
-                .join("rule-agents")
-                .join("sessions.json")
-        };
-
-        let mut sessions = SessionStore::new(session_file);
-        sessions.load_sessions().await.unwrap_or_else(|e| {
-            eprintln!("Warning: Could not load sessions: {}", e);
-        });
-
         Ok(Ruler {
-            rule_engine: Arc::new(rule_engine),
+            rules,
             agents: HashMap::new(),
-            sessions: Arc::new(Mutex::new(sessions)),
             test_mode: is_test,
         })
     }
@@ -80,796 +65,29 @@ impl Ruler {
             .ok_or_else(|| anyhow::anyhow!("Agent {} not found", agent_id))
     }
 
-    pub async fn handle_waiting_state(&self, agent_id: &str, capture: &str) -> Result<()> {
-        self.get_agent(agent_id).await?; // Ensure agent exists
-        let rules = self.rule_engine.get_rules().await;
-        let action = decide_action(capture, &rules);
-
-        println!("Agent {}: Capture \"{}\" → {:?}", agent_id, capture, action);
-
-        // Save session state before executing action (for potential recovery)
-        self.save_current_session_state(agent_id, capture)
-            .await
-            .unwrap_or_else(|e| {
-                eprintln!("Warning: Could not save session state: {}", e);
-            });
-
-        // Execute action with graceful error handling
-        match self.execute_action(agent_id, action).await {
-            Ok(()) => Ok(()),
-            Err(e) => {
-                eprintln!("❌ Action execution failed for agent {}: {}", agent_id, e);
-
-                // Try to recover from the error
-                if let Err(recovery_error) = self.handle_interrupted_session(agent_id, &e).await {
-                    eprintln!("❌ Session recovery also failed: {}", recovery_error);
-                }
-
-                Err(e)
-            }
-        }
+    pub async fn get_rules(&self) -> Vec<CompiledRule> {
+        self.rules.read().await.clone()
     }
 
-    /// Execute an action based on the new ActionType system
-    async fn execute_action(&self, agent_id: &str, action: ActionType) -> Result<()> {
-        let agent = self.get_agent(agent_id).await?;
-
-        match action {
-            ActionType::SendKeys(keys) => {
-                println!("→ Sending keys to agent {}: {:?}", agent_id, keys);
-                self.send_keys_to_agent(agent, keys).await?;
-            }
-            ActionType::Workflow(workflow_name, args) => {
-                println!(
-                    "→ Executing workflow '{}' for agent {} with args: {:?}",
-                    workflow_name, agent_id, args
-                );
-                self.execute_workflow(agent, &workflow_name, args).await?;
-            }
-            ActionType::Legacy(cmd_kind, args) => {
-                // Handle legacy commands during transition period
-                println!(
-                    "→ Executing legacy command {:?} for agent {} with args: {:?}",
-                    cmd_kind, agent_id, args
-                );
-                self.send_command_to_agent(agent, cmd_kind, args).await?;
-            }
-        }
+    pub async fn reload_rules(&self, rules_path: &str) -> Result<()> {
+        let new_rules = crate::ruler::rule_loader::load_rules(std::path::Path::new(rules_path))?;
+        let mut rules_guard = self.rules.write().await;
+        *rules_guard = new_rules;
+        println!("✅ Rules reloaded successfully from {}", rules_path);
         Ok(())
     }
 
-    /// Send keys directly to the terminal
-    async fn send_keys_to_agent(&self, agent: &Agent, keys: Vec<String>) -> Result<()> {
-        if self.test_mode {
-            println!(
-                "ℹ️ Test mode: would send keys {:?} to agent {}",
-                keys,
-                agent.id()
-            );
-            return Ok(());
-        }
-
-        for key in keys {
-            println!("  → Sending key: '{}'", key);
-            agent
-                .send_keys(&key)
-                .await
-                .map_err(|e| anyhow::anyhow!("Failed to send key '{}': {}", key, e))?;
-            // Small delay between keys to avoid overwhelming the terminal
-            tokio::time::sleep(Duration::from_millis(50)).await;
-        }
-
-        Ok(())
-    }
-
-    /// Execute a workflow by name
-    async fn execute_workflow(
-        &self,
-        agent: &Agent,
-        workflow_name: &str,
-        args: Vec<String>,
-    ) -> Result<()> {
-        match workflow_name {
-            "github_issue_resolution" => {
-                // Use the existing entry command logic for GitHub issue resolution
-                self.execute_entry_command(agent, args).await?;
-            }
-            _ => {
-                return Err(anyhow::anyhow!("Unknown workflow: {}", workflow_name));
-            }
-        }
-        Ok(())
-    }
-
-    async fn send_command_to_agent(
-        &self,
-        agent: &Agent,
-        command: CmdKind,
-        args: Vec<String>,
-    ) -> Result<()> {
-        match command {
-            CmdKind::Entry => {
-                println!(
-                    "→ Executing entry for agent {} with args: {:?}",
-                    agent.id(),
-                    args
-                );
-                self.execute_entry_command(agent, args).await?;
-            }
-            CmdKind::Resume => {
-                println!("→ Sending resume to agent {}", agent.id());
-                self.execute_resume_command(agent).await?;
-            }
-        }
-
-        Ok(())
-    }
-
-    async fn execute_entry_command(&self, agent: &Agent, args: Vec<String>) -> Result<()> {
-        // Extract issue number from args or parse from agent_id
-        let issue_number = if !args.is_empty() {
-            args[0].clone()
-        } else {
-            // Try to extract from agent_id if it contains issue number
-            agent.id().split('-').next_back().unwrap_or("1").to_string()
-        };
-
-        println!("🚀 Starting entry workflow for issue #{}", issue_number);
-
-        // Step 1: Git operations
-        self.handle_git_operations(agent, &issue_number).await?;
-
-        // Step 2: Create and switch to worktree
-        self.setup_worktree(agent, &issue_number).await?;
-
-        // Step 3: Create draft PR
-        self.create_draft_pr(agent, &issue_number).await?;
-
-        // Step 4: Implementation phase (this would integrate with actual implementation logic)
-        self.coordinate_implementation(agent, &issue_number).await?;
-
-        println!("✅ Entry workflow completed for issue #{}", issue_number);
-        Ok(())
-    }
-
-    async fn execute_resume_command(&self, agent: &Agent) -> Result<()> {
-        println!("▶️ Executing resume command for agent {}", agent.id());
-
-        // In test mode, just simulate resume
-        if self.test_mode {
-            println!("ℹ️ Test mode: resume command simulated");
-            return Ok(());
-        }
-
-        let session_state = {
-            let sessions = self.sessions.lock().await;
-            sessions.get_latest_session_for_agent(agent.id()).cloned()
-        };
-
-        // Try to find the most recent session for this agent
-        if let Some(session_state) = session_state {
-            println!("📄 Found session state for agent {}", agent.id());
-            println!("  - Working directory: {}", session_state.working_directory);
-            println!("  - Last command: {:?}", session_state.last_command);
-            println!("  - Session timestamp: {}", session_state.timestamp);
-
-            // Restore session state
-            self.restore_session_state(agent, &session_state).await?;
-
-            println!("✅ Session restored successfully for agent {}", agent.id());
-        } else {
-            println!("⚠️ No previous session found for agent {}", agent.id());
-            println!("   Starting fresh terminal session");
-
-            // Start a fresh session since no previous state exists
-            self.start_fresh_session(agent).await?;
-        }
-
-        Ok(())
-    }
-
-    async fn handle_git_operations(&self, agent: &Agent, issue_number: &str) -> Result<()> {
-        println!("📦 Handling git operations for issue #{}", issue_number);
-
-        // Skip actual git operations in test environment
-        if self.test_mode {
-            println!("ℹ️ Skipping git operations in test environment");
-            return Ok(());
-        }
-
-        // Check if we're in a worktree first
-        let pwd_result = agent
-            .execute_command("pwd")
-            .await
-            .map_err(|e| anyhow::anyhow!("Failed to get current directory: {}", e))?;
-
-        let current_dir = pwd_result.output.trim();
-
-        // If we're already in a worktree for this issue, we don't need to do git operations
-        if current_dir.contains(&format!(".worktree/issue-{}", issue_number)) {
-            println!("ℹ️ Already in correct worktree, skipping git operations");
-            return Ok(());
-        }
-
-        // Check if we're in the main worktree directory and need to navigate elsewhere
-        if current_dir.contains(".worktree") {
-            println!("ℹ️ In worktree environment, git operations handled by worktree setup");
-            return Ok(());
-        }
-
-        // Check current branch first
-        let branch_result = agent
-            .execute_command("git branch --show-current")
-            .await
-            .map_err(|e| anyhow::anyhow!("Failed to get current branch: {}", e))?;
-
-        let current_branch = branch_result.output.trim();
-
-        // Only checkout main if we're not already on it and not in a worktree
-        if current_branch != "main" {
-            let result = agent
-                .execute_command("git checkout main")
-                .await
-                .map_err(|e| anyhow::anyhow!("Failed to checkout main: {}", e))?;
-
-            if result.exit_code != Some(0) {
-                return Err(anyhow::anyhow!(
-                    "Git checkout main failed: {}",
-                    result.error
-                ));
-            }
-        }
-
-        let result = agent
-            .execute_command("git pull origin main")
-            .await
-            .map_err(|e| anyhow::anyhow!("Failed to pull main: {}", e))?;
-
-        if result.exit_code != Some(0) {
-            return Err(anyhow::anyhow!("Git pull failed: {}", result.error));
-        }
-
-        println!("✅ Git operations completed successfully");
-        Ok(())
-    }
-
-    async fn setup_worktree(&self, agent: &Agent, issue_number: &str) -> Result<()> {
-        println!("🌳 Setting up worktree for issue #{}", issue_number);
-
-        // Skip actual worktree operations in test environment
-        if self.test_mode {
-            println!("ℹ️ Skipping worktree operations in test environment");
-            return Ok(());
-        }
-
-        // Check if worktree already exists
-        let worktree_path = format!(".worktree/issue-{}", issue_number);
-        let check_result = agent
-            .execute_command(&format!("test -d {}", worktree_path))
-            .await;
-
-        if check_result.is_ok() && check_result.unwrap().exit_code == Some(0) {
-            println!("ℹ️ Worktree already exists, switching to it");
-            agent
-                .set_working_directory(&worktree_path)
-                .await
-                .map_err(|e| anyhow::anyhow!("Failed to switch to worktree: {}", e))?;
-        } else {
-            // Check if branch already exists first
-            let branch_name = format!("issue-{}", issue_number);
-            let branch_check = agent
-                .execute_command(&format!(
-                    "git show-ref --verify --quiet refs/heads/{}",
-                    branch_name
-                ))
-                .await;
-
-            let cmd = if branch_check.is_ok() && branch_check.unwrap().exit_code == Some(0) {
-                // Branch exists, add worktree without creating new branch
-                format!("git worktree add {} {}", worktree_path, branch_name)
-            } else {
-                // Branch doesn't exist, create new branch
-                format!("git worktree add {} -b {}", worktree_path, branch_name)
-            };
-
-            let result = agent
-                .execute_command(&cmd)
-                .await
-                .map_err(|e| anyhow::anyhow!("Failed to create worktree: {}", e))?;
-
-            if result.exit_code != Some(0) {
-                return Err(anyhow::anyhow!(
-                    "Worktree creation failed: {}",
-                    result.error
-                ));
-            }
-
-            agent
-                .set_working_directory(&worktree_path)
-                .await
-                .map_err(|e| anyhow::anyhow!("Failed to switch to worktree: {}", e))?;
-        }
-
-        println!("✅ Worktree setup completed");
-        Ok(())
-    }
-
-    async fn create_draft_pr(&self, agent: &Agent, issue_number: &str) -> Result<()> {
-        println!("📝 Creating draft PR for issue #{}", issue_number);
-
-        // Skip actual PR operations in test environment
-        if self.test_mode {
-            println!("ℹ️ Skipping PR operations in test environment");
-            return Ok(());
-        }
-
-        // Get issue details
-        let issue_cmd = format!("gh issue view {}", issue_number);
-        let issue_result = agent
-            .execute_command(&issue_cmd)
-            .await
-            .map_err(|e| anyhow::anyhow!("Failed to get issue details: {}", e))?;
-
-        if issue_result.exit_code != Some(0) {
-            return Err(anyhow::anyhow!(
-                "Failed to retrieve issue: {}",
-                issue_result.error
-            ));
-        }
-
-        // Extract title from issue output (simplified - would need proper parsing)
-        let title = format!("Fix issue #{}", issue_number);
-
-        // Create empty commit
-        let commit_cmd = format!("git commit --allow-empty -m \"{}\"", title);
-        let commit_result = agent
-            .execute_command(&commit_cmd)
-            .await
-            .map_err(|e| anyhow::anyhow!("Failed to create empty commit: {}", e))?;
-
-        if commit_result.exit_code != Some(0) {
-            return Err(anyhow::anyhow!(
-                "Empty commit failed: {}",
-                commit_result.error
-            ));
-        }
-
-        // Push branch
-        let branch_name = format!("issue-{}", issue_number);
-        let push_cmd = format!("git push -u origin {}", branch_name);
-        let push_result = agent
-            .execute_command(&push_cmd)
-            .await
-            .map_err(|e| anyhow::anyhow!("Failed to push branch: {}", e))?;
-
-        if push_result.exit_code != Some(0) {
-            return Err(anyhow::anyhow!("Branch push failed: {}", push_result.error));
-        }
-
-        // Create draft PR
-        let pr_body = format!("Closes #{}\n\n## Summary\nImplementation for issue #{}\n\n🤖 Generated with Terminal Backend", issue_number, issue_number);
-        let pr_cmd = format!(
-            "gh pr create --draft --title \"{}\" --body \"{}\"",
-            title, pr_body
-        );
-
-        let pr_result = agent
-            .execute_command(&pr_cmd)
-            .await
-            .map_err(|e| anyhow::anyhow!("Failed to create PR: {}", e))?;
-
-        if pr_result.exit_code != Some(0) {
-            return Err(anyhow::anyhow!("PR creation failed: {}", pr_result.error));
-        }
-
-        println!("✅ Draft PR created successfully");
-        println!("📎 PR URL: {}", pr_result.output.trim());
-        Ok(())
-    }
-
-    async fn coordinate_implementation(&self, agent: &Agent, issue_number: &str) -> Result<()> {
-        println!("🔧 Coordinating implementation for issue #{}", issue_number);
-
-        // This is where the actual implementation logic would be coordinated
-        // For now, this is a placeholder that demonstrates the integration point
-
-        // Example: Run tests
-        self.run_quality_checks(agent).await?;
-
-        println!("✅ Implementation coordination completed");
-        Ok(())
-    }
-
-    async fn run_quality_checks(&self, agent: &Agent) -> Result<()> {
-        println!("🧪 Running quality checks");
-
-        // Skip actual quality checks in test environment
-        if self.test_mode {
-            println!("ℹ️ Skipping quality checks in test environment");
-            return Ok(());
-        }
-
-        // Check if we have Cargo.toml (Rust project)
-        let cargo_check = agent.execute_command("test -f Cargo.toml").await;
-
-        if cargo_check.is_ok() && cargo_check.unwrap().exit_code == Some(0) {
-            // Run Rust quality checks
-            let checks = vec!["cargo fmt --check", "cargo clippy", "cargo test"];
-
-            for check in checks {
-                println!("Running: {}", check);
-                let result = agent
-                    .execute_command(check)
-                    .await
-                    .map_err(|e| anyhow::anyhow!("Failed to run {}: {}", check, e))?;
-
-                if result.exit_code != Some(0) {
-                    println!("⚠️ Check failed: {}", check);
-                    println!("Error: {}", result.error);
-                    // Continue with other checks but log the failure
-                }
-            }
-        }
-
-        println!("✅ Quality checks completed");
-        Ok(())
-    }
-
-    pub async fn manage_editor_session(
-        &self,
-        agent_id: &str,
-        file_path: &str,
-        editor: Option<&str>,
-    ) -> Result<()> {
-        let agent = self.get_agent(agent_id).await?;
-        let editor_cmd = editor.unwrap_or("vim");
-
-        println!("📝 Starting editor session: {} {}", editor_cmd, file_path);
-
-        // For HT backend, this would work interactively
-        // For direct backend, this would fail as expected
-        match agent.backend_type() {
-            "ht" => {
-                let cmd = format!("{} {}", editor_cmd, file_path);
-                let result = agent
-                    .execute_command(&cmd)
-                    .await
-                    .map_err(|e| anyhow::anyhow!("Failed to start editor: {}", e))?;
-
-                println!(
-                    "Editor session completed with exit code: {:?}",
-                    result.exit_code
-                );
-            }
-            _ => {
-                println!("ℹ️ Agent cannot handle interactive editor sessions");
-                println!("📁 File path for manual editing: {}", file_path);
-            }
-        }
-
-        Ok(())
-    }
-
-    async fn save_current_session_state(&self, agent_id: &str, last_command: &str) -> Result<()> {
-        if self.test_mode {
-            return Ok(());
-        }
-
-        let agent = self.get_agent(agent_id).await?;
-
-        // Get current terminal snapshot
-        let snapshot = agent
-            .take_snapshot()
-            .await
-            .map_err(|e| anyhow::anyhow!("Failed to take terminal snapshot: {}", e))?;
-
-        // Get current environment variables
-        let env_vars = agent
-            .get_environment()
-            .await
-            .map_err(|e| anyhow::anyhow!("Failed to get environment variables: {}", e))?;
-
-        // Get current working directory (simplified - in real implementation would be more sophisticated)
-        let working_directory = env_vars.get("PWD").unwrap_or(&"/tmp".to_string()).clone();
-
-        let mut sessions = self.sessions.lock().await;
-        sessions
-            .save_session_state(
-                agent_id,
-                &working_directory,
-                snapshot,
-                env_vars,
-                Some(last_command.to_string()),
-            )
-            .await?;
-
-        Ok(())
-    }
-
-    async fn restore_session_state(
-        &self,
-        agent: &Agent,
-        session_state: &crate::ruler::session::SessionState,
-    ) -> Result<()> {
-        println!("🔄 Restoring session state for agent {}", agent.id());
-
-        // Restore working directory
-        println!(
-            "📁 Restoring working directory: {}",
-            session_state.working_directory
-        );
-        agent
-            .set_working_directory(&session_state.working_directory)
-            .await
-            .map_err(|e| anyhow::anyhow!("Failed to restore working directory: {}", e))?;
-
-        // Restore environment variables (for HT backend, this would involve setting them)
-        for (key, value) in &session_state.environment_vars {
-            if key != "PWD" && key != "HOME" && key != "USER" {
-                // Skip system variables
-                let env_cmd = format!("export {}={}", key, value);
-                agent.send_keys(&env_cmd).await.ok(); // Best effort
-                agent.send_keys("\r").await.ok();
-            }
-        }
-
-        // If there was a last command, show it as a suggestion but don't execute it
-        if let Some(ref last_cmd) = session_state.last_command {
-            println!("💡 Last command was: {}", last_cmd);
-            println!("   You can press Up arrow or retype to continue");
-        }
-
-        // Show current terminal content summary
-        let content_lines = session_state.terminal_snapshot.content.lines().count();
-        println!(
-            "📺 Restored terminal session with {} lines of content",
-            content_lines
-        );
-
-        Ok(())
-    }
-
-    async fn start_fresh_session(&self, agent: &Agent) -> Result<()> {
-        println!("🆕 Starting fresh session for agent {}", agent.id());
-
-        // Clear terminal and show welcome message
-        agent.send_keys("clear").await.ok();
-        agent.send_keys("\r").await.ok();
-
-        tokio::time::sleep(tokio::time::Duration::from_millis(500)).await;
-
-        let welcome_msg = format!("echo '🤖 Agent {} session started'", agent.id());
-        agent.send_keys(&welcome_msg).await.ok();
-        agent.send_keys("\r").await.ok();
-
-        println!("✅ Fresh session initialized for agent {}", agent.id());
-        Ok(())
-    }
-
-    pub async fn cleanup_old_sessions(&self, max_age_hours: u64) -> Result<()> {
-        let mut sessions = self.sessions.lock().await;
-        sessions.cleanup_old_sessions(max_age_hours).await
-    }
-
-    pub async fn list_sessions(&self) -> Vec<crate::ruler::session::SessionState> {
-        let sessions = self.sessions.lock().await;
-        sessions.list_sessions().into_iter().cloned().collect()
-    }
-
-    pub async fn remove_session(&self, agent_id: &str) -> Result<bool> {
-        let mut sessions = self.sessions.lock().await;
-        sessions.remove_session(agent_id).await
-    }
-
-    async fn handle_interrupted_session(
-        &self,
-        agent_id: &str,
-        error: &anyhow::Error,
-    ) -> Result<()> {
-        println!(
-            "🔧 Attempting to recover from interrupted session for agent {}",
-            agent_id
-        );
-
-        if self.test_mode {
-            println!("ℹ️ Test mode: simulating session recovery");
-            return Ok(());
-        }
-
-        let agent = self.get_agent(agent_id).await?;
-
-        // Check if backend is still available
-        if !agent.is_available().await {
-            return Err(anyhow::anyhow!("Agent is no longer available"));
-        }
-
-        // Try to send Ctrl+C to cancel any hanging processes
-        println!("🛑 Sending interrupt signal...");
-        if let Err(e) = agent.send_keys("^C").await {
-            eprintln!("Failed to send interrupt: {}", e);
-        }
-
-        // Wait a moment for processes to terminate
-        tokio::time::sleep(tokio::time::Duration::from_millis(1000)).await;
-
-        // Try to reset the terminal state
-        println!("🔄 Resetting terminal state...");
-
-        // Send a few newlines to get to a clean prompt
-        for _ in 0..3 {
-            if let Err(e) = agent.send_keys("\r").await {
-                eprintln!("Failed to send newline: {}", e);
-            }
-            tokio::time::sleep(tokio::time::Duration::from_millis(200)).await;
-        }
-
-        // Try to execute a simple command to verify the terminal is responsive
-        println!("🧪 Testing terminal responsiveness...");
-        if agent.send_keys("echo 'Terminal recovered'").await.is_ok() {
-            agent.send_keys("\r").await.ok();
-            tokio::time::sleep(tokio::time::Duration::from_millis(500)).await;
-
-            // Take a snapshot to see if we got output
-            if let Ok(snapshot) = agent.take_snapshot().await {
-                if snapshot.content.contains("Terminal recovered") {
-                    println!("✅ Terminal session recovered successfully");
-
-                    // Save the recovery state
-                    self.save_current_session_state(agent_id, "session_recovered")
-                        .await
-                        .ok();
-                    return Ok(());
-                }
-            }
-        }
-
-        // If we can't recover, mark this in the session
-        println!("⚠️ Terminal recovery incomplete - session may need manual intervention");
-
-        // Update session state to indicate it needs attention
-        let mut sessions = self.sessions.lock().await;
-        if let Some(session) = sessions.get_session_mut(agent_id) {
-            session.last_command = Some(format!("ERROR: {}", error));
-        }
-        drop(sessions);
-
-        Err(anyhow::anyhow!("Terminal session recovery incomplete"))
-    }
-
-    pub async fn health_check(&self) -> Result<bool> {
-        if self.test_mode {
-            return Ok(true);
-        }
-
-        // Check health of all agents
-        for (agent_id, agent) in &self.agents {
-            if !agent.is_available().await {
-                println!("⚠️ Agent {} is not available", agent_id);
-                return Ok(false);
-            }
-
-            // Try a simple command to test responsiveness
-            match agent.send_keys("echo health_check").await {
-                Ok(()) => {
-                    agent.send_keys("\r").await.ok();
-                    tokio::time::sleep(tokio::time::Duration::from_millis(500)).await;
-
-                    // Check if we can take a snapshot
-                    match agent.take_snapshot().await {
-                        Ok(_) => continue,
-                        Err(_) => {
-                            println!("⚠️ Agent {} snapshot failed", agent_id);
-                            return Ok(false);
-                        }
-                    }
-                }
-                Err(_) => {
-                    println!("⚠️ Agent {} is not responsive", agent_id);
-                    return Ok(false);
-                }
-            }
-        }
-
-        Ok(true)
-    }
-
-    pub async fn force_cleanup_agent(&self, agent_id: &str) -> Result<()> {
-        println!("🧹 Force cleaning up agent session: {}", agent_id);
-
-        if !self.test_mode {
-            if let Ok(agent) = self.get_agent(agent_id).await {
-                // Send multiple interrupt signals
-                for _ in 0..3 {
-                    agent.send_keys("^C").await.ok();
-                    tokio::time::sleep(tokio::time::Duration::from_millis(100)).await;
-                }
-
-                // Clear the terminal
-                agent.send_keys("clear").await.ok();
-                agent.send_keys("\r").await.ok();
-            }
-        }
-
-        // Remove the session from persistence
-        self.remove_session(agent_id).await.ok();
-
-        println!("✅ Agent {} cleaned up", agent_id);
-        Ok(())
-    }
-
-    pub async fn emergency_stop_all(&self) -> Result<()> {
-        println!("🚨 Emergency stop: cleaning up all sessions");
-
-        if !self.test_mode {
-            for (agent_id, agent) in &self.agents {
-                // Send multiple interrupt signals
-                for _ in 0..5 {
-                    agent.send_keys("^C").await.ok();
-                    tokio::time::sleep(tokio::time::Duration::from_millis(100)).await;
-                }
-
-                // Try to cleanup the agent
-                agent.cleanup().await.ok();
-                println!("🛑 Stopped agent {}", agent_id);
-            }
-        }
-
-        // Clear all sessions
-        let mut sessions = self.sessions.lock().await;
-        sessions.clear_all_sessions();
-        sessions.save_sessions().await.ok();
-
-        println!("🛑 Emergency stop completed");
-        Ok(())
+    pub async fn decide_action_for_capture(&self, capture: &str) -> ActionType {
+        let rules = self.get_rules().await;
+        decide_action(capture, &rules)
     }
 }
 
 impl std::fmt::Debug for Ruler {
     fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
         f.debug_struct("Ruler")
-            .field("rule_engine", &self.rule_engine)
             .field("agents", &self.agents.keys().collect::<Vec<_>>())
             .field("test_mode", &self.test_mode)
             .finish()
     }
-}
-
-#[allow(dead_code)]
-pub trait AgentInterface {
-    fn send_command(
-        &self,
-        command: CmdKind,
-        args: Vec<String>,
-    ) -> impl std::future::Future<Output = AgentResult> + Send;
-}
-
-#[derive(Debug)]
-#[allow(dead_code)]
-pub enum AgentResult {
-    Success,
-    Retry,
-    Failed(String),
-}
-
-#[allow(dead_code)]
-pub async fn retry_with_backoff<F, Fut>(mut operation: F, max_attempts: u32) -> Result<()>
-where
-    F: FnMut() -> Fut,
-    Fut: std::future::Future<Output = Result<AgentResult>>,
-{
-    for attempt in 1..=max_attempts {
-        match operation().await? {
-            AgentResult::Success => return Ok(()),
-            AgentResult::Failed(err) => return Err(anyhow::anyhow!("Agent failed: {}", err)),
-            AgentResult::Retry => {
-                if attempt < max_attempts {
-                    let delay_ms = 100 * 2_u64.pow(attempt - 1);
-                    tokio::time::sleep(std::time::Duration::from_millis(delay_ms)).await;
-                } else {
-                    return Err(anyhow::anyhow!("Max retry attempts reached"));
-                }
-            }
-        }
-    }
-    Ok(())
 }
