--- src/ruler/rule_engine/compiled_rule.rs
@@ -1,5 +1,0 @@
-// This file is now integrated into rule_file.rs
-// CompiledRule and related structures are defined there
-// This file is kept for module structure compatibility
-
-pub use super::rule_file::CompiledRule;
