+++ src/workflow/error_handling.rs
@@ -1,0 +1,44 @@
+use anyhow::Result;
+
+#[allow(dead_code)]
+pub struct ErrorHandler;
+
+#[allow(dead_code)]
+impl ErrorHandler {
+    pub fn new() -> Self {
+        Self
+    }
+
+    pub fn handle_workflow_error(&self, error: &anyhow::Error) -> Result<()> {
+        eprintln!("🚨 Workflow error occurred: {}", error);
+
+        // Log error details for debugging
+        let error_chain = error
+            .chain()
+            .map(|e| e.to_string())
+            .collect::<Vec<_>>()
+            .join(" → ");
+
+        eprintln!("Error chain: {}", error_chain);
+
+        // Could implement different recovery strategies based on error type
+        // For now, just propagate the error
+        Err(anyhow::anyhow!("Workflow error: {}", error))
+    }
+
+    pub fn handle_execution_error(&self, error: &anyhow::Error) -> Result<()> {
+        eprintln!("⚠️ Execution error occurred: {}", error);
+
+        // Could implement retry logic or fallback strategies
+        // For now, just propagate the error
+        Err(anyhow::anyhow!("Execution error: {}", error))
+    }
+
+    pub fn handle_session_error(&self, error: &anyhow::Error) -> Result<()> {
+        eprintln!("📄 Session error occurred: {}", error);
+
+        // Could implement session recovery strategies
+        // For now, just propagate the error
+        Err(anyhow::anyhow!("Session error: {}", error))
+    }
+}
